+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 **Methods**
  - `isString`, `notString`
-
+ - `isArray`, `notArray`
 
 
 ## [7.1.0] - 2021-03-20